Add support for hardware monitoring (hwmon) devices
Add preliminary support for hardware monitoring (hwmon) devices.

The hwmon devices can be used exactly as if they were IIO devices. The
only thing to be careful about, is that the type of the channel is one
of the hwmon_chan_type enum instead of the iio_chan_type enum.

To avoid confusion a hwmon_channel_get_type() wrapper has been added as
well, which just returns the iio_channel's type as a hwmon_chan_type.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>
Co-developed-by: Robin Getz <robin.getz@analog.com>
Signed-off-by: Robin Getz <robin.getz@analog.com>